add test for jitted version of Book
