+++ tree/dataframe/test/CMakeLists.txt
@@ -22,7 +22,9 @@ ROOT_ADD_GTEST(dataframe_splitcoll_arrayview dataframe_splitcoll_arrayview.cxx T
 ROOT_ADD_GTEST(dataframe_redefine dataframe_redefine.cxx LIBRARIES ROOTDataFrame)
 target_include_directories(dataframe_splitcoll_arrayview PRIVATE ${CMAKE_CURRENT_SOURCE_DIR})
 if(NOT MSVC OR win_broken_tests)
-  ROOT_ADD_GTEST(dataframe_simple dataframe_simple.cxx LIBRARIES ROOTDataFrame)
+  ROOT_GENERATE_DICTIONARY(MaxSlotHelperDict MaxSlotHelper.h LINKDEF MaxSlotHelperLinkDef.h OPTIONS -inlineInputHeader)
+  ROOT_ADD_GTEST(dataframe_simple dataframe_simple.cxx MaxSlotHelperDict.cxx LIBRARIES ROOTDataFrame)
+  target_include_directories(dataframe_simple PRIVATE ${CMAKE_CURRENT_SOURCE_DIR})
   ROOT_ADD_GTEST(dataframe_snapshot dataframe_snapshot.cxx LIBRARIES ROOTDataFrame)
   ROOT_ADD_GTEST(dataframe_helpers dataframe_helpers.cxx LIBRARIES ROOTDataFrame)
   ROOT_ADD_GTEST(dataframe_vecops dataframe_vecops.cxx LIBRARIES ROOTDataFrame)
