+++ tree/dataframe/test/MaxSlotHelperLinkDef.h
@@ -1,0 +1,5 @@
+#ifdef __ROOTCLING__
+
+#pragma link C++ class MaxSlotHelper-;
+
+#endif
